subclassing FitnessTask - all tests OK
